+++ src/vstboard/views/configdialog.cpp
@@ -118,7 +118,7 @@ ConfigDialog::ConfigDialog(MainHost *myHost, QWidget *parent) :
 
 
     ui->bufferSize->addItem("Auto",0);
-    index = ui->bufferSize->findData( myHost->GetBufferSize() );
+    index = ui->bufferSize->findData( (unsigned int)myHost->GetBufferSize() );
     if(index==-1) {
         debug("ConfigDialog invalid buffer size")
         return;
